Refact/modulize-appjs: Moved elements and event listners into own modules.
