--- public/scripts/bundle.js.LICENSE.txt
@@ -1,1 +1,0 @@
-/*! regenerator-runtime -- Copyright (c) 2014-present, Facebook, Inc. -- license (MIT): https://github.com/facebook/regenerator/blob/main/LICENSE */
