+++ src/js/classes/Notes.js
