+++ src/js/classes/Storage.js
