+++ src/js/modules/elements.js
@@ -1,0 +1,6 @@
+// Form Elements
+export const $title = document.querySelector("#form-main #title");
+export const $body = document.querySelector("#form-main #body");
+
+// Color tooltip
+export const $colorToolTip = document.querySelector(".color-tooltip");
