+++ src/js/modules/load.js
@@ -1,0 +1,9 @@
+import UI from "../classes/ui";
+
+export const domContentLoaded = () => {
+  const uI = new UI();
+
+  document.addEventListener("DOMContentLoaded", () => {
+    uI.renderNotes();
+  });
+};
